refactor(brands): extract relocation service
