+++ apps/server/api/src/collections/brands/brands.module.ts
@@ -9,6 +9,7 @@ import { BrandsController } from '@api/collections/brands/controllers/brands.con
 import { BrandsRelationshipsController } from '@api/collections/brands/controllers/relationships/brands-relationships.controller';
 import { BrandDataMapper } from '@api/collections/brands/services/brand-data.mapper';
 import { BrandPersistenceService } from '@api/collections/brands/services/brand-persistence.service';
+import { BrandRelocationService } from '@api/collections/brands/services/brand-relocation.service';
 import { BrandSetupService } from '@api/collections/brands/services/brand-setup.service';
 import { BrandsService } from '@api/collections/brands/services/brands.service';
 import { DefaultRecurringContentService } from '@api/collections/brands/services/default-recurring-content.service';
@@ -70,6 +71,7 @@ import { forwardRef, Module } from '@nestjs/common';
     // OnboardingModule ↔ BrandsModule import cycle.
     BrandSetupService,
     BrandPersistenceService,
+    BrandRelocationService,
     BrandDataMapper,
     MasterPromptGeneratorService,
     CreditsGuard,
